some changes to reduce time and memory usage
The changes were intended to have three effects.
First, the initial SVD on the large matrix was
being done once for the pseudo-inverse and once
for the fiedler vector.  Now it is done only once.
Second, the laplacian square root that was being
subdivided during tree reconstruction had a constant
number of columns.  Now the number of columns is
never more than the number of rows.
Third, the memory usage was constant during tree building.
These changes were intended to reduce memory usage
during tree building, but the changes actually make it
increase.  This problem will be addressed in subsequent changes.